BateshSniperBoat/requirements.txt
requests
pandas
python-telegram-bot
bs4
yfinance
schedule
kiteconnect
nsepython